feat: add support for changing the theme of the new tab page
This commit adds support for changing the theme of the new tab page. The
theme can be changed by clicking on the theme button on the bottom right
corner of the page. The button opens a modal that displays a list of
available themes. When a theme is selected, the page background color
changes to the selected theme color.

The following files were modified:
- server.ts: change port variable case from lowercase port to uppercase
  PORT to improve semantics. Add support for process.env.PORT environment
  variable to be able to run app on a configurable port.
- package.json: add react-icons dependency.
- theme-modal.tsx: add ThemeModal component that displays a modal with a
  list of available themes. The modal is opened by clicking on the theme
  button on the bottom right corner of the page.
- open-button.tsx: add OpenButton component that displays the theme

feat(theme.ts): add module for theme with types for theme object and theme name types to be used in the app