feat(theme): add theme functionality to new tab page
- Add support for process.env.PORT environment variable to be able to run app on a configurable port
- Change port variable case from lowercase port to uppercase PORT to improve semantics
- Move open button component to theme folder
- Add theme container component to wrap dynamic background and open button components
- Add label component to display theme name and color
- Add theme labels list component to display all available themes
- Add theme modal component to display theme labels list and allow user to change theme
- Remove console.log statements from code

feat(theme-modal.tsx): add ThemeModal component to allow users to change the theme of the app. The component receives props to control its visibility, the current theme, the user name and a function to set the new theme. It also renders a list of available themes and a close button.